% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{florentzian}
\alias{florentzian}
\title{Calculates a Lorentzian bell centered at mean, with max height and a given  FWHM}
\usage{
florentzian(xx, mean, max, fwhm)
}
\arguments{
\item{xx}{The range of x values to compute the function (domain)}

\item{mean}{The center of the Lorentzian}

\item{max}{The max height of the curve}

\item{fwhm}{The full width at half maximum of the shape. It is not the same as the SD}
}
\value{
A numeric vector of length equal to length(xx)
}
\description{
Calculates a Lorentzian bell centered at mean, with max height and a given  FWHM
}
