% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{fpseudoVoigt}
\alias{fpseudoVoigt}
\title{Calculates a pseudoVoigt bell centered at mean, with max height and a given  FWHM.
A pseudoVoigt is an approximation of the Voigt function(Convolution of a Gaussian and a Lorentzian)
as a linear combination of the 2 aforementioned functions.}
\usage{
fpseudoVoigt(xx, mu, mean, max, fwhm)
}
\arguments{
\item{xx}{The range of x values to compute the function (domain)}

\item{mu}{The linear combination factor:  \code{mu*lorentzian + (1-mu)*gaussian}}

\item{mean}{The center of the Voigt}

\item{max}{The max height of the curve}

\item{fwhm}{The full width at half maximum of the shape. It is not the same as the SD}
}
\value{
A numeric vector of length equal to length(xx)
}
\description{
Calculates a pseudoVoigt bell centered at mean, with max height and a given  FWHM.
A pseudoVoigt is an approximation of the Voigt function(Convolution of a Gaussian and a Lorentzian)
as a linear combination of the 2 aforementioned functions.
}
