% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shiftSignal}
\alias{shiftSignal}
\title{Change the chemical shift of a \code{\linkS4class{NMRSignal1D}}}
\usage{
shiftSignal(signal, to, by = 0, hertz = FALSE, SF = 600)
}
\arguments{
\item{signal, }{\code{\linkS4class{NMRSignal1D}}}

\item{to, }{numeric, optional, new chemical shift coordinate}

\item{by, }{numeric, optional, shift to apply}

\item{hertz, }{logic, is \code{by} given in Hz (TRUE), or in chemical shift units
(FALSE, defauilt)?}

\item{SF, }{spectrometer field, in MHz. Only used if \code{hertz=TRUE}. Default 600}
}
\value{
\code{\linkS4class{NMRSignal1D}}, the shifted signal
}
\description{
Change the chemical shift of a \code{\linkS4class{NMRSignal1D}}
}
\details{
If \code{to} is provided, the signal is shifted to the given chemical shift.
Otherwise, if \code{by} is provided, the signal is shifted by the given ppm.
Else, the signal is left unshifted
}
