% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulateSignal}
\alias{simulateSignal}
\title{Simulate a signal from resonance parameters (c.shift and j coupling). No second order effects.}
\usage{
simulateSignal(
  cshift,
  j = NULL,
  multiplicity = NULL,
  frequency = 600,
  linewidth = 1,
  mu = 0.85,
  intensity = 1,
  nbAtoms = 1,
  fwhm
)
}
\arguments{
\item{cshift, }{numeric, chemical shift. Compulsory argument.}

\item{j, }{numeric, coupling constants (Hz). The default NULL produces a singlet}

\item{multiplicity, }{character. Use letters s,d,t,q,p,x,h,o,n for
singlet, doublet, triplet, etc. If omitted will be estimated from j, see details}

\item{frequency, }{numeric, default 600 (MHz)}

\item{linewidth, }{numeric, specified as full-width at half-maximum. Default 1 (Hz)}

\item{mu, }{numeric, gaussian-to-lorentzian ratio in the pseudo-voigt
combination. Default 0.85}

\item{intensity, }{numeric, intensity of the signal (summit). Default 1}

\item{nbAtoms, }{numeric, number of resonating atoms. Default 1}

\item{fwhm, }{numeric, optional. Quality of life alias for \code{linewidth}.
The corresponding NMRSignalModel1D parameter in this library is called fwhm,
which may confuse veterans. This parameter allows them to auto-pilot when
calling this function.}
}
\value{
a \code{\linkS4class{NMRSignal1D}}
}
\description{
Simulate a signal from resonance parameters (c.shift and j coupling). No second order effects.
}
\details{
This is rather a class constructor on a different signature; I could not
figure out the S4 way of doing it.

The signal can be specified following customary NMR practice i.e. passing
a coupling constant for each group of magnetically equivalent nuclei,
along with the corresponding multiplicities. Then, the function also accepts
more unusual method of giving a j for each individual proton. For instance,
\code{j=c(3,2),multiplicity="td"} produces the same outcome as \code{j=c(3,3,2)}.

Be warned: internally, the function runs very few checks on the arguments
and does its best to produce a reasonable output.
}
