% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scaleSignal}
\alias{scaleSignal}
\title{Scale the intensity of a \code{\linkS4class{NMRSignal1D}}}
\usage{
scaleSignal(signal, to, by = 1)
}
\arguments{
\item{signal, }{\code{\linkS4class{NMRSignal1D}}}

\item{to, }{numeric, optional, new maximum intensity}

\item{by, }{optional, height scaling factor}
}
\value{
\code{\linkS4class{NMRSignal1D}}, the scaled signal
}
\description{
Scale the intensity of a \code{\linkS4class{NMRSignal1D}}
}
\details{
If \code{to} is provided, the signal is scaled so that its maximum
intensity matches the parameter. Otherwise, if \code{by} is provided, the
signal is scaled by the given factor. Else, the signal is left unscaled
}
